LayoutTests/http/tests/webgpu/webgpu/compat/api/validation/createBindGroupLayout-expected.txt

CONSOLE MESSAGE: Unhandled Promise Rejection: SyntaxError: Importing binding name 'kCompatModeUnsupportedStorageTextureFormats' is not found.

Harness Error (FAIL), message = Unhandled rejection: Importing binding name 'kCompatModeUnsupportedStorageTextureFormats' is not found.
